Fix mistake in target pdf
A fidelitous implementation would have implemented the log likelihood with the log of a factorial of x.
That gave overflows for x ~60-100, so the log gamma function was used as the gamma function is the factorial function for integers.
The error was that the gamma function at zero is undetermined and implemented as inf.